refactor(api): split the four remaining oversized handler files
Same package, pure moves with their doc comments, one concern per file:

- failover.go (929 -> 654) keeps the handler, the response types, the
  group CRUD, Candidates and GetByModelUUID;
  failover_circuitbreaker.go takes the four CB interfaces,
  CircuitBreakerStatus with its cache and the reset endpoints.
- discovery.go (885 -> 563) keeps registration, the status/ack
  endpoints and discovery itself; discovery_quota.go takes serveQuota,
  the usage/balance/Ollama-account endpoints and RefreshAllQuotas;
  discovery_dismiss.go takes the dismiss and unpin claim endpoints.
- stats.go (843 -> 240) keeps the handler, the response types, the
  param parsing, GetStats and calculateStats; stats_queries.go takes
  the per-dimension queries; stats_timeseries.go takes GetTimeSeries,
  fillEmptyBuckets and GetProviderDistribution.
- logs.go (809 -> 425) keeps the entry types and the four endpoints;
  logs_query.go takes the select columns, query building, filters,
  keyset predicate, cursor encode/decode, scanning and sort defs.

All four entries leave the size allowlist. Tests are unchanged.
