+++ internal/api/discovery.go
@@ -2,8 +2,6 @@ package api
 
 import (
 	"context"
-	"encoding/json"
-	"errors"
 	"fmt"
 	"net/http"
 	"time"
@@ -19,7 +17,6 @@ import (
 	"github.com/hugalafutro/model-hotel/internal/failover"
 	"github.com/hugalafutro/model-hotel/internal/model"
 	"github.com/hugalafutro/model-hotel/internal/provider"
-	"github.com/hugalafutro/model-hotel/internal/quota"
 	"github.com/hugalafutro/model-hotel/internal/util"
 )
 
@@ -365,129 +362,6 @@ func (h *Handler) DiscoverProviderModels(w http.ResponseWriter, r *http.Request)
 	writeJSON(w, response)
 }
 
-// respondQuotaError maps a quota/usage/balance fetch error to an HTTP response.
-// A dead upstream credential (provider.ErrProviderKeyInvalid) is a provider
-// configuration problem, not a server fault: respond 424 Failed Dependency - a
-// 4xx, so the access log records it at WARN rather than ERROR - and skip the
-// ERROR line respondError would emit, since the fetch layer already logged the
-// rejection once at WARN. The sidebar badge hides on any non-2xx, so the dead
-// provider simply shows no badge instead of spamming errors on every poll.
-// Anything else stays a logged 500.
-func respondQuotaError(w http.ResponseWriter, providerName, resource string, err error) {
-	if errors.Is(err, provider.ErrProviderKeyInvalid) {
-		http.Error(w, fmt.Sprintf("provider key invalid or inactive for %s", providerName), http.StatusFailedDependency)
-		return
-	}
-	respondError(w, fmt.Sprintf("failed to fetch %s for provider %s", resource, providerName), err, http.StatusInternalServerError)
-}
-
-// serveQuota returns the stored quota snapshot for prov, reproducing its status.
-// On a cold miss it performs a one-time live fetch, persists it, then serves it,
-// so a brand-new provider's first view is not blank. The X-Quota-Fetched-At
-// header (RFC3339) carries the snapshot age for the client display.
-func (h *Handler) serveQuota(w http.ResponseWriter, r *http.Request, prov *provider.Provider, expectedKind string) {
-	providerType := provider.TypeOf(prov)
-	kind, ok := quotaKindFor(providerType)
-	if !ok || kind != expectedKind {
-		// Enforce the endpoint contract: /usage, /balance and /account each
-		// serve only their own kind. Without this a provider whose type maps to
-		// a different kind would return the wrong payload shape on the endpoint.
-		http.Error(w, expectedKind+" information not supported for this provider type", http.StatusBadRequest)
-		return
-	}
-
-	snap, err := h.quotaRepo.Get(r.Context(), prov.ID, kind)
-	if err != nil {
-		respondError(w, "failed to load quota snapshot", err, http.StatusInternalServerError)
-		return
-	}
-	if snap == nil {
-		// Cold: fetch once and persist, then fall through to serve it. The
-		// context is decoupled from the HTTP request deadline so a client
-		// disconnect does not abort the in-flight upstream call mid-fetch.
-		ctx, cancel := context.WithTimeout(context.WithoutCancel(r.Context()), 30*time.Second)
-		defer cancel()
-		disc := newDiscoveryService()
-		k, payload, status, ferr := fetchQuotaSnapshot(ctx, disc, prov, h.cfg.MasterKey)
-		if ferr != nil {
-			respondQuotaError(w, prov.Name, kind, ferr)
-			return
-		}
-		if uerr := h.quotaRepo.Upsert(ctx, quota.Snapshot{ProviderID: prov.ID, Kind: k, Payload: payload, HTTPStatus: status, Source: "manual"}); uerr != nil {
-			respondError(w, "failed to persist quota snapshot", uerr, http.StatusInternalServerError)
-			return
-		}
-		snap, _ = h.quotaRepo.Get(ctx, prov.ID, kind)
-		if snap == nil {
-			http.Error(w, "quota unavailable", http.StatusServiceUnavailable)
-			return
-		}
-	}
-
-	w.Header().Set("X-Quota-Fetched-At", snap.FetchedAt.UTC().Format(time.RFC3339))
-	switch snap.HTTPStatus {
-	case http.StatusNoContent:
-		w.WriteHeader(http.StatusNoContent)
-	case http.StatusFailedDependency:
-		http.Error(w, "", http.StatusFailedDependency)
-	default:
-		w.Header().Set("Content-Type", "application/json")
-		//nolint:gosec // G705 false positive: provider quota JSON body, not HTML; Content-Type is application/json
-		_, _ = w.Write(snap.Payload)
-	}
-}
-
-// GetProviderUsage serves usage/quota information for a provider from the
-// read-through snapshot store (cold-filling on first view).
-func (h *Handler) GetProviderUsage(w http.ResponseWriter, r *http.Request) {
-	providerID, ok := parseUUIDParam(w, r, "id", "provider ID")
-	if !ok {
-		return
-	}
-
-	prov, err := h.providerRepo.Get(r.Context(), providerID)
-	if err != nil {
-		respondLookupError(w, err, pgx.ErrNoRows, "provider not found", "failed to load provider")
-		return
-	}
-
-	h.serveQuota(w, r, prov, "usage")
-}
-
-// GetProviderBalance serves balance information for a provider from the
-// read-through snapshot store (cold-filling on first view).
-func (h *Handler) GetProviderBalance(w http.ResponseWriter, r *http.Request) {
-	providerID, ok := parseUUIDParam(w, r, "id", "provider ID")
-	if !ok {
-		return
-	}
-
-	prov, err := h.providerRepo.Get(r.Context(), providerID)
-	if err != nil {
-		respondLookupError(w, err, pgx.ErrNoRows, "provider not found", "failed to load provider")
-		return
-	}
-
-	h.serveQuota(w, r, prov, "balance")
-}
-
-// GetOllamaCloudAccount serves Ollama Cloud account info from the read-through
-// snapshot store (cold-filling on first view).
-func (h *Handler) GetOllamaCloudAccount(w http.ResponseWriter, r *http.Request) {
-	providerID, ok := parseUUIDParam(w, r, "id", "provider ID")
-	if !ok {
-		return
-	}
-
-	prov, err := h.providerRepo.Get(r.Context(), providerID)
-	if err != nil {
-		respondLookupError(w, err, pgx.ErrNoRows, "provider not found", "failed to load provider")
-		return
-	}
-
-	h.serveQuota(w, r, prov, "account")
-}
-
 // DiscoverAllResult holds the result of discovering models from a single provider.
 type DiscoverAllResult struct {
 	ProviderName string         `json:"provider_name"`
@@ -686,200 +560,4 @@ func (h *Handler) discoverAllProviders(ctx context.Context, recordMisses bool) (
 	}
 
 	return results, succeeded, failed, totalDiscovered, nil
-}
-
-// QuotaRefreshResult holds the result of refreshing quotas for a single provider.
-type QuotaRefreshResult struct {
-	ProviderName string `json:"provider_name"`
-	ProviderType string `json:"provider_type"`
-	Refreshed    bool   `json:"refreshed"`
-	Error        string `json:"error,omitempty"`
-}
-
-// RefreshAllQuotas refreshes quota information for all providers that support it.
-func (h *Handler) RefreshAllQuotas(w http.ResponseWriter, r *http.Request) {
-	providers, err := h.providerRepo.List(r.Context())
-	if err != nil {
-		respondError(w, "failed to list providers", nil, http.StatusInternalServerError)
-		return
-	}
-
-	discovery := newDiscoveryService()
-
-	var results []QuotaRefreshResult
-	refreshed := 0
-	failed := 0
-	skipped := 0
-
-	for _, prov := range providers {
-		if !prov.Enabled {
-			continue
-		}
-
-		provCtx, provCancel := context.WithTimeout(context.Background(), 30*time.Second)
-
-		providerType := provider.TypeOf(prov)
-		result := QuotaRefreshResult{
-			ProviderName: prov.Name,
-			ProviderType: providerType,
-		}
-
-		// fetchQuotaSnapshot returns kind=="" for provider types that expose no
-		// quota endpoint; those are skipped (and not added to results), matching
-		// the prior behaviour where the type switch had no matching case.
-		kind, payload, status, ferr := fetchQuotaSnapshot(provCtx, discovery, prov, h.cfg.MasterKey)
-		if kind == "" {
-			provCancel()
-			skipped++
-			continue
-		}
-		if ferr != nil {
-			result.Error = ferr.Error()
-			failed++
-			_ = h.quotaRepo.RecordFailure(provCtx, prov.ID, kind, ferr.Error())
-		} else if uerr := h.quotaRepo.Upsert(provCtx, quota.Snapshot{ProviderID: prov.ID, Kind: kind, Payload: payload, HTTPStatus: status, Source: "manual"}); uerr != nil {
-			// A failed persist must not report success: the stored snapshot is
-			// still stale, so surface it as a failure rather than "refreshed".
-			debuglog.Warn("quota: manual refresh upsert failed", "provider", prov.Name, "error", uerr)
-			result.Error = uerr.Error()
-			failed++
-		} else {
-			result.Refreshed = true
-			refreshed++
-		}
-
-		provCancel()
-		results = append(results, result)
-	}
-
-	writeJSON(w, map[string]any{
-		"results":   results,
-		"refreshed": refreshed,
-		"failed":    failed,
-		"skipped":   skipped,
-	})
-}
-
-// DismissDiscoveryClaimsRequest carries the models to dismiss on one provider.
-//
-// Dismiss-only, deliberately: there is no un-dismiss direction. A dismissal
-// self-heals, because models.Upsert clears discovery_dismissed_at on any
-// sighting, so the next discovery run undoes it for any model that came back.
-// That is the only reversal the feature needs, and it needs no endpoint.
-//
-// A traffic-retired model gets there by a different route, since Upsert
-// deliberately preserves its dismissal (it is sighted on every scan, so clearing
-// on a sighting would make it impossible to silence). For those, the operator
-// enabling the model clears the dismissal in the same statement as the enable
-// (models.SetEnabled and models.Update), so a model retired again afterwards
-// raises a fresh claim. That has to be atomic rather than left to the next
-// sighting: traffic reaches a re-enabled model in seconds and a scan is about an
-// hour away, so a second retirement would otherwise arrive first and re-arm the
-// preserve-the-dismissal rule around a stamp nothing could clear. Still no
-// endpoint needed.
-type DismissDiscoveryClaimsRequest struct {
-	ProviderID string   `json:"provider_id"`
-	ModelIDs   []string `json:"model_ids"`
-}
-
-// DismissDiscoveryClaims stamps the operator dismissal for models on one
-// provider. setModelsDismissed only touches rows that are currently
-// enabled=false and not manually disabled, so a suspect (still enabled) or
-// healthy model cannot be pre-dismissed; those affect zero rows and fall
-// through the 404 path below like any other unmatched model ID.
-//
-// Deliberately NOT added to httpx.IsReadOnlyExemptPost: unlike the discovery-change
-// ack it sits beside, this suppresses a real discrepancy from every operator's
-// view, which is a genuine state change.
-func (h *Handler) DismissDiscoveryClaims(w http.ResponseWriter, r *http.Request) {
-	var req DismissDiscoveryClaimsRequest
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		http.Error(w, "invalid request body", http.StatusBadRequest)
-		return
-	}
-	providerID, err := uuid.Parse(req.ProviderID)
-	if err != nil {
-		http.Error(w, "invalid provider ID", http.StatusBadRequest)
-		return
-	}
-	if len(req.ModelIDs) == 0 {
-		http.Error(w, "model_ids must not be empty", http.StatusBadRequest)
-		return
-	}
-
-	dismissed, err := setModelsDismissed(r.Context(), h.dbPool.Pool(), providerID, req.ModelIDs)
-	if err != nil {
-		respondError(w, "failed to dismiss discovery claims", err, http.StatusInternalServerError)
-		return
-	}
-	if len(dismissed) == 0 {
-		http.Error(w, "no matching models", http.StatusNotFound)
-		return
-	}
-
-	// `dismissed` names the models actually stamped, so a partial result is fully
-	// informative: the caller marks exactly those and leaves the rest alone.
-	// `updated` is kept for compatibility and is simply its length.
-	writeJSON(w, map[string]any{"dismissed": dismissed, "updated": len(dismissed)})
-}
-
-// UnpinDiscoveryClaimsRequest carries the models to unpin on one provider. It is
-// shaped exactly like DismissDiscoveryClaimsRequest because it is the same kind
-// of operation from the modal's side: a bulk verdict on a provider's rows.
-//
-// Unpin-only, like dismiss, and for a matching reason: the pin direction is not
-// an endpoint. A pin is armed by the operator enabling the model
-// (models.SetEnabled and models.Update stamp manually_enabled_at), and cleared
-// automatically by the next sighting, since a listed model needs no exemption
-// from the listing. This endpoint covers the one case neither of those reaches:
-// the operator changing their mind about a model the provider still does not
-// list, where there is nothing to enable and no sighting coming.
-type UnpinDiscoveryClaimsRequest struct {
-	ProviderID string   `json:"provider_id"`
-	ModelIDs   []string `json:"model_ids"`
-}
-
-// UnpinDiscoveryClaims drops the operator pin from models on one provider,
-// returning them to discovery's listing-based auto-disable with a clean
-// miss-streak. setModelsUnpinned only touches rows that actually carry a pin, so
-// an already-unpinned model affects zero rows and falls through the 404 path
-// below like any other unmatched model ID.
-//
-// No model cache invalidation: the pin and the miss-streak live only in the
-// database. model.Model carries neither, so no cached entry can go stale on this
-// write, and the dismiss endpoint beside it invalidates nothing for the same
-// reason. What does change on the next scan — the model being disabled — flows
-// through the same path any auto-disable does.
-//
-// Deliberately NOT added to httpx.IsReadOnlyExemptPost: it hands a model back to
-// automatic management, which is a genuine state change.
-func (h *Handler) UnpinDiscoveryClaims(w http.ResponseWriter, r *http.Request) {
-	var req UnpinDiscoveryClaimsRequest
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		http.Error(w, "invalid request body", http.StatusBadRequest)
-		return
-	}
-	providerID, err := uuid.Parse(req.ProviderID)
-	if err != nil {
-		http.Error(w, "invalid provider ID", http.StatusBadRequest)
-		return
-	}
-	if len(req.ModelIDs) == 0 {
-		http.Error(w, "model_ids must not be empty", http.StatusBadRequest)
-		return
-	}
-
-	unpinned, err := setModelsUnpinned(r.Context(), h.dbPool.Pool(), providerID, req.ModelIDs)
-	if err != nil {
-		respondError(w, "failed to unpin discovery claims", err, http.StatusInternalServerError)
-		return
-	}
-	if len(unpinned) == 0 {
-		http.Error(w, "no matching models", http.StatusNotFound)
-		return
-	}
-
-	// `unpinned` names the rows actually cleared, so a partial result is fully
-	// informative: the caller updates exactly those and leaves the rest alone.
-	writeJSON(w, map[string]any{"unpinned": unpinned})
 }
