+++ scripts/ci/size-allowlist.txt
@@ -9,11 +9,7 @@
 # Format: path<TAB>lines
 cmd/server/main.go	860
 internal/api/applogs_test.go	2083
-internal/api/discovery.go	885
 internal/api/discovery_integration_quota_test.go	2735
-internal/api/failover.go	929
-internal/api/logs.go	809
-internal/api/stats.go	843
 internal/failover/circuitbreaker_test.go	2056
 internal/frontdesk/autosync_test.go	4254
 internal/model/model_crud_test.go	2399
